+++ types/health.go
@@ -1,8 +1,0 @@
-package types
-
-type ClusterHealth struct {
-	MemberID uint64
-	Name     string
-	Err      error
-	Healthy  bool
-}