+++ types/status_map.go
@@ -13,3 +13,15 @@ func (s StatusMap) Healthy() bool {
 	}
 	return true
 }
+
+// ClusterHealth holds cluster member status info.
+type ClusterHealth struct {
+	// MemberID is the etcd cluster member's ID.
+	MemberID uint64
+	// Name is the cluster member's name.
+	Name string
+	// Err holds any errors encountered while checking the member's health.
+	Err error
+	// Healthy describes the health of the cluster member.
+	Healthy bool
+}