+++ CHANGELOG.md
@@ -22,6 +22,7 @@ enterprise edition.
 - Refactoring of the cmd package for sensuctl to allow easier customization in
 the enterprise edition.
 - Upgrade dep to v0.5.0
+- Added cluster health information to /health endpoint in sensu-backend.
 
 ### Fixed
 - Fixed `sensuctl completion` help for bash and zsh.